requirements.txt
anykeystore==0.2
apex==0.9.10.dev0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cryptacular==1.6.2
defusedxml==0.7.1
greenlet==3.1.1
hupper==1.12.1
idna @ file:///croot/idna_1714398848350/work
joblib==1.4.2
MarkupSafe==2.1.5
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
PasteDeploy==3.1.0
pbkdf2==1.3
pillow @ file:///croot/pillow_1721059439630/work
plaster==1.1.2
plaster-pastedeploy==1.0.1
protobuf==5.29.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python3-openid==3.2.0
pytorch-ignite==0.4.1
PyYAML==6.0.2
repoze.sendmail==4.4.1
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
scikit-learn==1.3.2
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.38
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
transaction==5.0
translationstring==1.4
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1727769808118/work
velruse==1.1.1
venusian==3.1.1
WebOb==1.8.9
WTForms==3.1.2
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==5.0
zope.interface==7.2
zope.sqlalchemy==3.1
